Rename & move BaseRemoteEvaluator to DelegateEvaluator in evaluation.
