+++ fedot/core/optimisers/optimizer.py
@@ -13,7 +13,7 @@ from fedot.core.optimisers.gp_comp.operators.operator import PopulationT
 from fedot.core.optimisers.graph import OptGraph
 from fedot.core.optimisers.objective import GraphFunction, Objective, ObjectiveFunction
 from fedot.core.optimisers.opt_node_factory import DefaultOptNodeFactory, OptNodeFactory
-from fedot.remote.base_remote_evaluator import BaseRemoteEvaluator
+from fedot.core.optimisers.gp_comp.evaluation import DelegateEvaluator
 
 OptimisationCallback = Callable[[PopulationT, GenerationKeeper], Any]
 
@@ -56,13 +56,13 @@ class GraphGenerationParams:
     verifier: GraphVerifier
     advisor: DefaultChangeAdvisor
     node_factory: OptNodeFactory
-    remote_evaluator: Optional[BaseRemoteEvaluator] = None
+    remote_evaluator: Optional[DelegateEvaluator] = None
 
     def __init__(self, adapter: Optional[BaseOptimizationAdapter] = None,
                  rules_for_constraint: Sequence[VerifierRuleType] = (),
                  advisor: Optional[DefaultChangeAdvisor] = None,
                  node_factory: Optional[OptNodeFactory] = None,
-                 remote_evaluator: Optional[BaseRemoteEvaluator] = None,
+                 remote_evaluator: Optional[DelegateEvaluator] = None,
                  ):
         self.adapter = adapter or IdentityAdapter()
         self.verifier = GraphVerifier(rules_for_constraint, self.adapter)
